+++ openbb_terminal/helper_funcs.py
@@ -47,7 +47,6 @@ from openbb_terminal.core.config.paths import (
     USER_EXPORTS_DIRECTORY,
 )
 from openbb_terminal.core.config import paths
-from openbb_terminal import keys_model
 
 logger = logging.getLogger(__name__)
 
@@ -1804,18 +1803,3 @@ def list_from_str(value: str) -> List[str]:
     if value:
         return value.split(",")
     return []
-
-
-def first_time_user() -> bool:
-    """Whether a user is a first time user. A first time user is someone with an empty .env file.
-    If this is true, it also adds an env variable to make sure this does not run again.
-
-    Returns
-    ----------
-    bool
-        Whether or not the user is a first time user
-    """
-    if USER_ENV_FILE.stat().st_size == 0:
-        keys_model.set_key("OPENBB_PREVIOUS_USE", "True", True)
-        return True
-    return False
